Added relationship queries to the show handler.
